Revert "Accounting version2"
