--- accounting-ledger-application/.idea/.name
@@ -1,1 +1,0 @@
-Main.java
